chore: rename dog-registration-modal --> dog-profile-modal
